requirements.txt
# -------- pytorch -------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.2.4
torchmetrics[image]==1.4.0
piq

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# ---------- cv ----------- #
pillow
opencv-python
scikit-image

# ---------- ds ----------- #
matplotlib
seaborn
pandas

# --------- unsplash --------- #
requests 
pillow
pyunsplash 
python-dotenv

# --------- loggers --------- #
wandb

# --------- others --------- #
rootutils       # standardizing the project root setup
autoroot        # set path to root directory
autorootcwd     # set path to root directory
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
black           # code formatter
isort           # import sorter
# sh            # for running bash commands in some tests (linux/macos only)
